tests: use fail point (#114)
* tests: use fail point

Remove the hacky function pointer, and use failpoint to do the precise
control.

* Add note why failpoints is seperate

* cargo fmt
